+++ saleor/graphql/account/mutations/account.py
@@ -1,11 +1,12 @@
 import graphene
-from django.core.exceptions import ValidationError
+from django.core.exceptions import ObjectDoesNotExist, ValidationError
 
 from ....account import emails, events as account_events, models, utils
 from ....checkout import AddressType
 from ...account.enums import AddressTypeEnum
 from ...account.types import Address, AddressInput, User
 from ...core.mutations import BaseMutation, ModelMutation
+from .base import send_user_password_reset_email
 from .staff import CustomerCreate, UserAddressInput
 
 
@@ -184,3 +185,33 @@ class CustomerSetDefaultAddress(BaseMutation):
 
         utils.change_user_default_address(user, address, address_type)
         return cls(user=user)
+
+
+class CustomerPasswordResetInput(graphene.InputObjectType):
+    email = graphene.String(
+        required=True,
+        description=("Email of the user that will be used for password recovery."),
+    )
+
+
+class CustomerPasswordReset(BaseMutation):
+    class Arguments:
+        input = CustomerPasswordResetInput(
+            description="Fields required to reset customer's password", required=True
+        )
+
+    class Meta:
+        description = (
+            "DEPRECATED: Use SetPassword instead. Resets the customer's password."
+        )
+
+    @classmethod
+    def perform_mutation(cls, _root, info, **data):
+        email = data.get("input")["email"]
+        try:
+            user = models.User.objects.get(email=email)
+        except ObjectDoesNotExist:
+            raise ValidationError({"email": "User with this email doesn't exist"})
+        site = info.context.site
+        send_user_password_reset_email(user, site)
+        return CustomerPasswordReset()
