+++ saleor/graphql/account/mutations/staff.py
@@ -1,7 +1,7 @@
 from copy import copy
 
 import graphene
-from django.core.exceptions import ObjectDoesNotExist, ValidationError
+from django.core.exceptions import ValidationError
 from graphql_jwt.decorators import staff_member_required
 from graphql_jwt.exceptions import PermissionDenied
 
@@ -308,25 +308,6 @@ class StaffDelete(StaffDeleteMixin, UserDelete):
         # ID so that the success response contains ID of the deleted object.
         instance.id = db_id
         return cls.success_response(instance)
-
-
-class PasswordReset(BaseMutation):
-    class Arguments:
-        email = graphene.String(description="Email", required=True)
-
-    class Meta:
-        description = "Sends password reset email"
-        permissions = ("account.manage_users",)
-
-    @classmethod
-    def perform_mutation(cls, _root, info, email):
-        try:
-            user = models.User.objects.get(email=email)
-        except ObjectDoesNotExist:
-            raise ValidationError({"email": "User with this email doesn't exist"})
-        site = info.context.site
-        send_user_password_reset_email(user, site)
-        return PasswordReset()
 
 
 class AddressCreate(ModelMutation):
