+++ saleor/graphql/account/schema.py
@@ -11,6 +11,7 @@ from .mutations.account import (
     AccountRequestDeletion,
     AccountUpdate,
     CustomerAddressCreate,
+    CustomerPasswordReset,
     CustomerRegister,
     CustomerSetDefaultAddress,
     LoggedUserUpdate,
@@ -18,7 +19,7 @@ from .mutations.account import (
 from .mutations.base import (
     AddressDelete,
     AddressUpdate,
-    CustomerPasswordReset,
+    PasswordReset,
     SetPassword,
     UserClearStoredMeta,
     UserUpdateMeta,
@@ -29,7 +30,6 @@ from .mutations.staff import (
     CustomerCreate,
     CustomerDelete,
     CustomerUpdate,
-    PasswordReset,
     StaffCreate,
     StaffDelete,
     StaffUpdate,
